feat: modalized get started
