+++ apps/extension/src/ui/domains/Portfolio/GetStarted/LearnMore/index.ts
@@ -1,0 +1,2 @@
+export * from "./useLearnMoreModal"
+export * from "./LearnMoreModal"
