+++ apps/extension/src/ui/domains/Portfolio/GetStarted/LearnMore/useLearnMoreModal.ts
@@ -1,0 +1,3 @@
+import { useGlobalOpenClose } from "@talisman/hooks/useGlobalOpenClose"
+
+export const useLearnMoreModal = () => useGlobalOpenClose("LearnMore")
