+++ apps/extension/src/ui/domains/Portfolio/GetStarted/TryTalisman/index.ts
@@ -1,0 +1,2 @@
+export * from "./useTryTalismanModal"
+export * from "./TryTalismanModal"
