+++ apps/extension/src/ui/domains/Portfolio/GetStarted/TryTalisman/useTryTalismanModal.ts
@@ -1,0 +1,3 @@
+import { useGlobalOpenClose } from "@talisman/hooks/useGlobalOpenClose"
+
+export const useTryTalismanModal = () => useGlobalOpenClose("TryTalisman")
